增加一个新用户

+++ src/main/java/manage/dao/DepartmentMapper.java
@@ -6,7 +6,7 @@ import manage.bean.DepartmentExample;
 import org.apache.ibatis.annotations.Param;
 
 public interface DepartmentMapper {
-    int countByExample(DepartmentExample example);
+    long countByExample(DepartmentExample example);
 
     int deleteByExample(DepartmentExample example);
